[Maven-Central] Changing groupd id using github repository group id.
