--- src/main/java/es/squdan/querydsl/filters/QueryDslFiltersException.java
@@ -1,16 +1,0 @@
-package es.squdan.querydsl.filters;
-
-/**
- * QueryDslFilter context exception.
- */
-public class QueryDslFiltersException extends RuntimeException {
-
-    public QueryDslFiltersException(final String msg) {
-        super(msg);
-    }
-
-    public QueryDslFiltersException(final Throwable exception, final String msg) {
-        super(msg, exception);
-    }
-
-}
