+++ src/main/java/io/github/squdan/querydsl/filters/QueryDslFiltersException.java
@@ -1,0 +1,16 @@
+package io.github.squdan.querydsl.filters;
+
+/**
+ * QueryDslFilter context exception.
+ */
+public class QueryDslFiltersException extends RuntimeException {
+
+    public QueryDslFiltersException(final String msg) {
+        super(msg);
+    }
+
+    public QueryDslFiltersException(final Throwable exception, final String msg) {
+        super(msg, exception);
+    }
+
+}
