--- src/test/java/es/squdan/querydsl/filters/repository/entity/BankAccountId.java
@@ -1,18 +1,0 @@
-package es.squdan.querydsl.filters.repository.entity;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@Embeddable
-@NoArgsConstructor
-@AllArgsConstructor
-public class BankAccountId {
-
-    private String bank;
-    private String account;
-}
