--- src/test/java/es/squdan/querydsl/filters/repository/entity/UserEntity.java
@@ -1,56 +1,0 @@
-package es.squdan.querydsl.filters.repository.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.SourceType;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-import java.util.List;
-import java.util.UUID;
-
-@Data
-@Entity
-@Builder
-@Table(name = "users")
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @Column(nullable = false, unique = true)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private Roles role;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column
-    private String lastName;
-
-    @Column
-    private BigDecimal savings;
-
-    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY, cascade = CascadeType.REMOVE)
-    private List<BankAccountEntity> accounts;
-
-    @CreationTimestamp(source = SourceType.DB)
-    private Instant createdOn;
-
-    @UpdateTimestamp(source = SourceType.DB)
-    private Instant lastUpdatedOn;
-}
