+++ src/test/java/io/github/squdan/querydsl/filters/configuration/DatabaseTestConfiguration.java
@@ -1,0 +1,11 @@
+package io.github.squdan.querydsl.filters.configuration;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EntityScan("io.github.squdan")
+@EnableJpaRepositories(basePackages = "io.github.squdan")
+public class DatabaseTestConfiguration {
+}
